requirements.txt
Django==5.1.6
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Faker==37.1.0
pillow==11.1.0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
celery==5.5.1
django-cors-headers==4.7.0
django-filter==25.1
psycopg2-binary==2.9.10
django-modeltranslation==0.19.13
